requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
slack-sdk==3.26.0
slack-bolt==1.18.0
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3 